Create lib folder and move files
