+++ lib/interfaces/Block.ts
