+++ pages/_app.tsx
@@ -1,5 +1,5 @@
 import type { AppProps } from 'next/app'
-import 'utils/tailwind.css'
+import 'lib/tailwind.css'
 
 function MyApp({ Component, pageProps }: AppProps) {
   return <Component {...pageProps} />
